-Removing support for microblaze backend, and replacing it with MIPS (implemented by Tim Grant).
-Completely reworked the assembler, which should be more amenable to debug info (compile with -g), although debug info still not fully used (only file:line:col info is used for now).


Former-commit-id: 1671632ed14c1e735b427ef2f78d10df6ef595f1

+++ llvm_trax/lib/trax_trax.cpp
@@ -11,12 +11,10 @@ extern int printf ( const char * format, ... )
 
   // Arguments will all be placed on the stack in subsequent words.
   // Just need to pass the address of the first arg to the intrinsic
-  trax_printformat((int)&format);
+  trax_printformat(&format);
 
   // Return something just to match the signature of stdio::printf
   // Ideally this would return the number of args successfully printed
   return 0;
 
 }
-
-
